Cleanup for Master release
