+++ Text_Files/Error.log
@@ -10,3 +10,4 @@ Error downloading: pError downloading: /Error downloading: rError downloading: /
 Error downloading: /program/FREP48000584
 Error downloading: /program/FOLA01006581
 Error downloading: /program/FREP48000584
+Error downloading: /program/FNYH78003978
